Function version_check::is_max_version

pub fn is_max_version(max_version: &str) -> Option<bool>

Checks that the running or installed rustc is at most some maximum version.

The format of max_version is a semantic version: 1.3.0, 1.15.0-beta, 1.14.0, 1.16.0-nightly, etc.

If the version cannot be retrieved or parsed, or if max_version could not be parsed, returns None. Otherwise returns true if the installed rustc is at most max_version and false otherwise.