Function version_check::triple

pub fn triple() -> Option<(Version, Channel, Date)>

Reads the triple of Version, Channel, and Date of the installed or running rustc.

If any attribute cannot be determined (see the top-level documentation), returns None.

To obtain only one of three attributes, use Version::read(), Channel::read(), or Date::read().